requirements.txt
paramiko==2.7.1
stepic==0.4.1
Pillow==7.1.1
requests==2.23.0
piexif==1.1.3
